moved register logic to authservice
